refactor(ai-gateway): split providers, consolidate streaming bodies
- Move per-provider logic into src/providers/{workers-ai,openai,anthropic,custom}.ts
  plus src/providers/shared.ts (resolveBaseUrl, cfGatewayHeaders,
  withTimeoutSignal, parseRawToolCalls). gateway.ts drops from 804 to 240 LOC.
- Extract CF Universal Endpoint body+normalize into src/fallback.ts.
- Streaming body builders now reuse buildAnthropicBody / buildOpenAiBody
  from the non-streaming path instead of duplicating message shaping.
- SSE parser accepts \r\n\r\n separators in addition to \n\n.
- Consumer-canceled streams now abort the upstream fetch via a per-request
  linked AbortController.

Public API unchanged. 167/167 tests pass (+2 new: CRLF parse, cancel→abort).
